Add files via upload
Twitter has become a major platform for satisfied or dissatisfied customers to reflect their opinion on their business experience. Customer reviews provide useful information that could assist the business to identify and resolve the existing issues. Natural language processing could assist the business in summarization, sorting and categorizing the reviews in a fraction of time. 
Archive of reviews exist can be used in topic modeling. These topics can be filtered to provide insight on the urgency of the issue. In addition, the established model can be used to process the new reviews in the same manner.